Working photon run with namedtuple args * example1.py, example1.ipynb to run this untested routine * fixed type issues for jit * Tried Numba cuda simulator without success * user's howfar and ausgab pushed into egsnrc.photon since can't be passed to kernel

+++ egsnrc/particles.py
@@ -42,12 +42,18 @@ def set_particle(i, regions, iparticles, fparticles):
 @device_jit
 def replace_e_uvw(p, energy, u, v, w):
     """Return a Particle like `p` but with uvw replaced"""
-    return Particle(p.status, p.region, energy, p.x, p.y, p.z, u, v, w)
+    return Particle(
+        p.status, p.region, np.float32(energy), p.x, p.y, p.z,
+        np.float32(u), np.float32(v), np.float32(w)
+    )
 
 @device_jit
 def replace_region_xyz(p, region, x, y, z):
     """Return a Particle like `p` but with x,y,z replaced"""
-    return Particle(p.status, region, p.energy, x, y, z, p.u, p.v, p.w)
+    return Particle(
+        p.status, region, p.energy,
+        np.float32(x), np.float32(y), np.float32(z), p.u, p.v, p.w
+    )
 
 
 class PhotonSource:
